+++ llvm/include/llvm/Support/Error.h
@@ -21,7 +21,6 @@
 #include "llvm/Support/ErrorHandling.h"
 #include "llvm/Support/ErrorOr.h"
 #include "llvm/Support/Format.h"
-#include "llvm/Support/FormatVariadic.h"
 #include "llvm/Support/raw_ostream.h"
 #include <cassert>
 #include <cstdint>
@@ -1334,21 +1333,6 @@ template <typename... Ts>
 inline Error createStringError(std::errc EC, char const *Fmt,
                                const Ts &... Vals) {
   return createStringError(std::make_error_code(EC), Fmt, Vals...);
-}
-
-// LLVM formatv versions of llvm::createStringError
-
-template <typename... Ts>
-inline Error createStringErrorV(std::error_code EC, char const *Fmt,
-                                Ts &&...Vals) {
-  return make_error<StringError>(formatv(Fmt, std::forward<Ts>(Vals)...).str(),
-                                 EC, true);
-}
-
-template <typename... Ts>
-inline Error createStringErrorV(char const *Fmt, Ts &&...Vals) {
-  return createStringErrorV(llvm::inconvertibleErrorCode(), Fmt,
-                            std::forward<Ts>(Vals)...);
 }
 
 /// This class wraps a filename and another Error.
